switch geometry structs to only contain the JSON Array
The geometry JSON objects generally only contain "type" and 
"coordinates". The type is already known since it is the name of the 
struct. Therefore we can just put the "coordinates" inside, which is a 
JSON3.Array. Doing some benchmarks on the previous geometry structs, 
which all needed to access the coordinates field, I found that simply 
accessing this field was quite slow. Implementing it this way avoids 
that cost. The downside is that "foreign members" that may be contained 
in the geometry object will be lost, however this is not required by the 
spec.