support bbox for geometries
Requiring this led to quite the refactor. Now the geometry structs don't only store the `JSON3.Array` with their coordinates, but the entire object. This includes the optional bbox, and any GeoJSON object may have other members according to the spec. The object is not restricted to a `JSON3.Object`, since these are the products of reading JSON strings. If we want to create another sub-geometry using `GeoInterface.getgeom` we need to be able to create them as well, For this we use a NamedTuple in the geometry, similar to the `_lower` approach for writing.